combine new-email-received & new-email-received-instant
