+++ components/gmail/actions/add-label-to-email/add-label-to-email.mjs
@@ -4,7 +4,7 @@ export default {
   key: "gmail-add-label-to-email",
   name: "Add Label to Email",
   description: "Add label(s) to an email message. [See the docs](https://developers.google.com/gmail/api/reference/rest/v1/users.messages/modify)",
-  version: "0.0.4",
+  version: "0.0.5",
   type: "action",
   props: {
     gmail,
