+++ components/gmail/actions/create-draft/create-draft.mjs
@@ -1,10 +1,11 @@
 import gmail from "../../gmail.app.mjs";
+import { ConfigurationError } from "@pipedream/platform";
 
 export default {
   key: "gmail-create-draft",
   name: "Create Draft",
   description: "Create a draft from your Google Workspace email account. [See the documentation](https://developers.google.com/gmail/api/reference/rest/v1/users.drafts/create)",
-  version: "0.0.2",
+  version: "0.0.3",
   type: "action",
   props: {
     gmail,
@@ -44,10 +45,16 @@ export default {
         "bodyType",
       ],
     },
-    attachments: {
+    attachmentFilename: {
       propDefinition: [
         gmail,
-        "attachments",
+        "attachmentFilename",
+      ],
+    },
+    attachmentUrlOrPath: {
+      propDefinition: [
+        gmail,
+        "attachmentUrlOrPath",
       ],
     },
     inReplyTo: {
@@ -67,6 +74,10 @@ export default {
     },
   },
   async run({ $ }) {
+    if ((this.attachmentFilename && ! this.attachmentUrlOrPath)
+      || (!this.attachmentFilename && this.attachmentUrlOrPath)) {
+      throw new ConfigurationError("Must specify both `Attachment Filename` and `Attachment URL or Path`");
+    }
     const opts = await this.gmail.getOptionsToSendEmail($, this);
     const response = await this.gmail.createDraft(opts);
     $.export("$summary", "Successfully created a draft message");
