+++ components/gmail/actions/find-email/find-email.mjs
@@ -4,7 +4,7 @@ export default {
   key: "gmail-find-email",
   name: "Find Email",
   description: "Find an email using Google's Search Engine. [See the docs](https://developers.google.com/gmail/api/reference/rest/v1/users.messages/list)",
-  version: "0.0.5",
+  version: "0.0.6",
   type: "action",
   props: {
     gmail,
