+++ components/gmail/actions/remove-label-from-email/remove-label-from-email.mjs
@@ -4,7 +4,7 @@ export default {
   key: "gmail-remove-label-from-email",
   name: "Remove Label from Email",
   description: "Remove label(s) from an email message. [See the docs](https://developers.google.com/gmail/api/reference/rest/v1/users.messages/modify)",
-  version: "0.0.1",
+  version: "0.0.2",
   type: "action",
   props: {
     gmail,
