+++ components/gmail/actions/send-email/send-email.mjs
@@ -1,10 +1,11 @@
 import gmail from "../../gmail.app.mjs";
+import { ConfigurationError } from "@pipedream/platform";
 
 export default {
   key: "gmail-send-email",
   name: "Send Email",
   description: "Send an email from your Google Workspace email account. [See the documentation](https://developers.google.com/gmail/api/reference/rest/v1/users.messages/send)",
-  version: "0.1.4",
+  version: "0.1.5",
   type: "action",
   props: {
     gmail,
@@ -32,6 +33,12 @@ export default {
         "fromName",
       ],
     },
+    fromEmail: {
+      propDefinition: [
+        gmail,
+        "fromEmail",
+      ],
+    },
     replyTo: {
       propDefinition: [
         gmail,
@@ -56,10 +63,16 @@ export default {
         "bodyType",
       ],
     },
-    attachments: {
+    attachmentFilename: {
       propDefinition: [
         gmail,
-        "attachments",
+        "attachmentFilename",
+      ],
+    },
+    attachmentUrlOrPath: {
+      propDefinition: [
+        gmail,
+        "attachmentUrlOrPath",
       ],
     },
     inReplyTo: {
@@ -76,6 +89,10 @@ export default {
     },
   },
   async run({ $ }) {
+    if ((this.attachmentFilename && ! this.attachmentUrlOrPath)
+      || (!this.attachmentFilename && this.attachmentUrlOrPath)) {
+      throw new ConfigurationError("Must specify both `Attachment Filename` and `Attachment URL or Path`");
+    }
     const opts = await this.gmail.getOptionsToSendEmail($, this);
     const response = await this.gmail.sendEmail(opts);
     $.export("$summary", `Successfully sent email to ${this.to}`);
