+++ components/gmail/actions/update-primary-signature/update-primary-signature.mjs
@@ -4,7 +4,7 @@ export default {
   key: "gmail-update-primary-signature",
   name: "Update Signature for Primary Email Address",
   description: "Update the signature for the primary email address. [See the documentation](https://developers.google.com/gmail/api/reference/rest/v1/users.settings.sendAs/update)",
-  version: "0.0.2",
+  version: "0.0.3",
   type: "action",
   props: {
     gmail,
