+++ components/gmail/sources/common/polling-history.mjs
@@ -6,6 +6,9 @@ export default {
   hooks: {
     ...common.hooks,
     async deploy() {
+      if (this.triggerType === "webhook") {
+        return;
+      }
       const historyId = await this.getHistoryId();
       if (!historyId) {
         return;
